move util to utils folder
