+++ src/sections/ProofOfExploit.tsx
@@ -11,10 +11,10 @@ import { Modal } from '@/components/Modal';
 import { Snackbar } from '@/components/Snackbar';
 import { NoData } from '@/components/ui/NoData';
 import { useGetFile } from '@/hooks/useFiles';
-import { useSearchParams } from '@/hooks/useSearchParams';
 import { copyToClipboard } from '@/utils/copyToClipboard.util';
 import { getDescription } from '@/utils/risk.util';
 import { StorageKey } from '@/utils/storage/useStorage.util';
+import { useSearchParams } from '@/utils/url.util';
 
 export function ProofOfExploit() {
   const { searchParams, removeSearchParams } = useSearchParams();
