+++ src/sections/detailsDrawer/SeedDrawer.tsx
@@ -11,12 +11,12 @@ import { OverflowText } from '@/components/OverflowText';
 import { Table } from '@/components/table/Table';
 import { DetailsListContainer } from '@/components/ui/DetailsListContainer';
 import { useGenericSearch } from '@/hooks/useGenericSearch';
-import { useSearchParams } from '@/hooks/useSearchParams';
 import { change as changeSeed } from '@/hooks/useSeeds';
 import { useSearchContext } from '@/state/search';
 import { formatDate } from '@/utils/date.util';
 import { getRoute } from '@/utils/route.util';
 import { StorageKey } from '@/utils/storage/useStorage.util';
+import { useSearchParams } from '@/utils/url.util';
 
 import { Asset, Attribute, Risk, Seed, SeedStatus } from '../../types';
